requirements.txt
apache-airflow[crypto]==1.9.0
matplotlib==2.2.2
pydruid==0.4.3
